chore: move chunk into chunks in storage
