+++ safenode/src/client/api.rs
@@ -14,10 +14,9 @@ use super::{
 use crate::{
     network::{NetworkEvent, SwarmDriver},
     protocol::{
-        chunk::Chunk,
         error::Error as ProtocolError,
         messages::{Cmd, CmdResponse, Query, QueryResponse, Request, Response},
-        storage::ChunkAddress,
+        storage::{Chunk, ChunkAddress},
     },
 };
