+++ safenode/src/protocol/storage/chunks/chunk.rs
@@ -6,7 +6,7 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use super::storage::ChunkAddress;
+use super::ChunkAddress;
 
 use bytes::Bytes;
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
