+++ safenode/src/protocol/storage/chunks/mod.rs
@@ -6,8 +6,11 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
+mod chunk;
+
+pub use self::chunk::Chunk;
+
 use crate::protocol::{
-    chunk::Chunk,
     error::{Error, Result},
     storage::ChunkAddress,
 };
